fixed compilation bug

+++ iot/models/condition.go
@@ -4,8 +4,8 @@ import "time"
 
 type Condition interface {
 	IsLightOn(time time.Time) bool
-	IsHeaterOn(time time.Time, temperature float32) bool
-	IdealTemperature(time time.Time) float32
+	IsHeaterOn(time time.Time, temperature float64) bool
+	IdealTemperature(time time.Time) float64
 }
 
 type ConditionBase struct {
